+++ tartufo/cli.py
@@ -49,7 +49,15 @@ class TartufoCLI(click.MultiCommand):
     "--rules",
     multiple=True,
     type=click.File("r"),
-    help="Path(s) to regex rules json list file(s).",
+    help="[DEPRECATED] Use the rule-patterns config options instead. Path(s) to regex "
+    "rules json list file(s).",
+)
+@click.option(
+    "--rule-patterns",
+    multiple=True,
+    type=str,
+    hidden=True,
+    help="Regular expression patterns to search the target for. May be specified multiple times.",
 )
 @click.option(
     "--default-regexes/--no-default-regexes",