+++ tartufo/types.py
@@ -36,6 +36,7 @@ class OutputFormat(enum.Enum):
 class GlobalOptions:
     __slots__ = (
         "rules",
+        "rule_patterns",
         "default_regexes",
         "entropy",
         "regex",
@@ -58,6 +59,7 @@ class GlobalOptions:
         "entropy_sensitivity",
     )
     rules: Tuple[TextIO, ...]
+    rule_patterns: Tuple[Dict[str, str], ...]
     default_regexes: bool
     entropy: bool
     regex: bool